refactor: filenames en kebab-case
